turning the map into a components, create the base nav component
progress!
