+++ package.json
@@ -16,6 +16,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-leaflet": "^3.2.0",
+    "react-router-dom": "^5.2.0",
     "typescript": "^4.1.2",
     "web-vitals": "^1.0.1"
   },
